# blog_system_backend/src/main/java/com/example/blog_system_backend/category/CategoryController.java
package com.example.blog_system_backend.category;

import com.example.blog_system_backend.category.dto.CategoryCreateRequest;
import com.example.blog_system_backend.category.dto.CategoryResponse;
import com.example.blog_system_backend.category.dto.CategoryUpdateRequest;
import jakarta.validation.Valid;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/categories")
public class CategoryController {

    private final CategoryService categoryService;

    public CategoryController(CategoryService categoryService) { this.categoryService = categoryService; }

    @GetMapping
    public List<CategoryResponse> list() { return categoryService.getAll(); }

    @GetMapping("/{id}")
    public CategoryResponse getById(@PathVariable Long id) {
        return categoryService.getById(id);
    }

    @PostMapping
    public CategoryResponse create(@Valid @RequestBody CategoryCreateRequest request) {
        return categoryService.create(request);
    }

    @PutMapping("/{id}")
    public CategoryResponse update(@PathVariable Long id, @Valid @RequestBody CategoryUpdateRequest request) {
        return categoryService.update(id, request);
    }

    @DeleteMapping("/{id}")
    public void delete(@PathVariable Long id) { categoryService.delete(id); }

}