# blog_system_frontend/doc/API-v2.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "包含用户注册登录、博客发布与管理、分类标签、评论及点赞功能。",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/users/me": {
      "get": {
        "summary": "获取当前登录用户信息",
        "deprecated": false,
        "description": "",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "summary": "更新当前用户资料",
        "deprecated": false,
        "description": "",
        "tags": [
          "users"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/{user_id}": {
      "get": {
        "summary": "查看指定用户公开信息",
        "deprecated": false,
        "description": "",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/users/me/blogs": {
      "get": {
        "summary": "查看当前用户的博客列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "blogs"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResult_Blog"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/{userId}/blogs": {
      "get": {
        "summary": "查看指定用户的博客列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "blogs"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "2",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResult_Blog"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/categories": {
      "get": {
        "summary": "获取全部分类列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "categories"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "summary": "创建分类(管理员、不含描述)",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "name": "美食分享"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/categories/{categoryId}": {
      "put": {
        "summary": "修改分类（管理员）",
        "deprecated": false,
        "description": "",
        "tags": [
          "categories"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "3",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              },
              "example": {
                "name": "CBA"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden",
            "description": "权限不足"
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "删除分类（需无博客关联，管理员）",
        "deprecated": false,
        "description": "",
        "tags": [
          "categories"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "3",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功",
            "headers": {}
          },
          "400": {
            "description": "仍有博客关联时删除失败",
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden",
            "description": "权限不足"
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/comments/{commentId}/replies": {
      "post": {
        "summary": "回复某条评论（仅限顶级评论）",
        "deprecated": false,
        "description": "",
        "tags": [
          "comments"
        ],
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "5",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              },
              "example": {
                "content": "忍住感觉"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "回复成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误（如试图回复子评论）",
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/comments/{commentId}": {
      "delete": {
        "summary": "删除评论（评论者、博客作者或管理员）",
        "deprecated": false,
        "description": "",
        "tags": [
          "comments"
        ],
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "7",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功",
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden",
            "description": "权限不足"
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/comments/{commentId}/like": {
      "post": {
        "summary": "点赞/取消点赞评论",
        "deprecated": false,
        "description": "",
        "tags": [
          "likes"
        ],
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "10",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "点赞成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "like_count": {
                      "type": "integer"
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "已点赞过",
            "headers": {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "未认证或 Token 无效"
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/comments": {
      "get": {
        "summary": "管理员获取全站评论列表",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/comments/{comment_id}": {
      "delete": {
        "summary": "管理员删除任意评论",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "23",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "avatar_url": {
            "type": "string",
            "format": "uri"
          },
          "bio": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string"
          },
          "avatar_url": {
            "type": "string",
            "format": "uri"
          },
          "bio": {
            "type": "string"
          }
        }
      },
      "Blog": {
        "type": "object",
        "properties": {
          "blog_id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "author": {
            "$ref": "#/components/schemas/User"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "view_count": {
            "type": "integer"
          },
          "like_count": {
            "type": "integer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CategoryRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "Tag": {
        "type": "object",
        "properties": {
          "tag_id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Comment": {
        "type": "object",
        "properties": {
          "comment_id": {
            "type": "integer",
            "format": "int64"
          },
          "blog_id": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "parent_comment_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "content": {
            "type": "string"
          },
          "like_count": {
            "type": "integer"
          },
          "replies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CommentRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
      "PageResult_Blog": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Blog"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "未认证或 Token 无效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "权限不足",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "servers": [],
  "security": [
    {
      "bearer": []
    }
  ]
}